% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_dsc.R
\name{eval_dsc}
\alias{eval_dsc}
\title{Evaluate Dynamic Subset Combination Results}
\usage{
eval_dsc(
  oos_y,
  oos_benchmark,
  oos_forecast_dsc,
  oos_dates,
  oos_chosen_parameter,
  model_names_tvc,
  oos_models_idx
)
}
\arguments{
\item{oos_y}{y A matrix of dimension `T * 1` or Numeric Vector of length `T`
containing the out-of-sample observations of the target variable.}

\item{oos_benchmark}{A matrix of dimension `T * 1` or
Numeric Vector of length `T` containing the
out-of-sample forecasts of a arbitrary benchmark.}

\item{oos_forecast_dsc}{A matrix of dimension `T * 1`
or Numeric Vector of length `T` containing the
out-of-sample forecasts of the Dynamic Subset Combination.}

\item{oos_dates}{A vector that contains all ouf-of-sample dates.}

\item{oos_chosen_parameter}{Matrix from 'dsc' that contains the
chosen (tuned) parameter from the Dynamic Subset Combination.}

\item{model_names_tvc}{Character Vector from [tvp()] that contains
all model names from the density forecasts.}

\item{oos_models_idx}{List from [dsc()] that contains
the column-indices of the selected density forecasts.}
}
\value{
List that contains
(1) the Clark-West-Statistic,
(2) the Out-of-Sample R2,
(3) the CSED between the DSC-Forecast and the benchmark,
(4) a boolean matrix that indicates when a predictor was chosen from the DSC
(5) and a CSED Plot.
}
\description{
`eval_dsc()` can be used to evaluate the results from [dsc()].
It provides statistical measures such as Clark-West-Statistic or the OOS-R2,
visualizes for the predictive performance as well as the selected parameters
and returns a matrix which indicates the selected predictors
at every point in time.
}
