% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{Ext_PF_CPILFESL}
\alias{Ext_PF_CPILFESL}
\title{External Point Forecasts for U.S. inflation (Core CPI)}
\format{
A \link{data.frame} with 245 quarterly observations (rows) and 16 forecasting models (columns):
\describe{
   \item{ARp}{AR(2) with intercept, estimated with OLS}
   \item{SBAR}{Bayesian Structural Breaks AR(2)}
   \item{UCSV}{Unobserved Components Stochastic Volatility Model}
   \item{TVPAR}{Time-Varying Parameter AR Model}
   \item{FAC5}{Principal Components Regression using the first five components}
   \item{BAG/FAC5}{Regression estimated using Bagging using the first five components}
   \item{GPR/FAC5}{Gaussian Process Regression using the first five components}
   \item{DMA/FAC5}{Dynamic Model Averaging using the first five components}
   \item{TVD/FAC5}{Time-Varying Dimension Model using the first five components}
   \item{VBDVS/FAC5}{Variational Bayes Dynamic Variable Selection Algorithm using the first five components}
   \item{SSVS/FAC60}{Stochastic Search Variable Selection algorithm using the first 60 components}
   \item{ELN/FAC60}{Elastic Net using the first 60 components}
   \item{VBDVS/FAC60}{Variational Bayes Dynamic Variable Selection Algorithm using the first 60 components}
   \item{ELN/X}{Elastic Net with all 439 predictors}
   \item{PLS/X}{Partial Least Squares with all 439 predictors}
   \item{VBDVS/X}{Variational Bayes Dynamic Variable Selection Algorithm with all 439 predictors}
}
}
\source{
\url{https://onlinelibrary.wiley.com/doi/full/10.1111/iere.12623}
}
\usage{
Ext_PF_CPILFESL
}
\description{
External point forecasts for quarterly Core CPI
              generated by the MatLab Code from Koop and Korobilis (2023).
              Out-of-sample one-step-ahead forecasts are generated
              from 1976-Q1 to 2021-Q4.
}
\references{
Koop, G. and Korobilis, D. (2023) "Bayesian dynamic variable selection in high dimensions."
\emph{International Economic Review}.
}
\keyword{datasets}
