% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdflex-package.R
\name{hdflex}
\alias{hdflex}
\alias{hdflex-package}
\title{hdflex: High-Dimensional Density Forecasts}
\description{
hdflex contains the forecasting algorithm STSC
developed by Ad채mmer, Lehmann and Sch체ssler (2023) <doi:10.2139/ssrn.4342487>.
STSC is a novel time series forecasting method designed to handle very large
sets of predictive signals, many of which are irrelevant or have only
short-lived predictive power.
Please cite the paper when using the package.
}
\author{
Philipp Ad채mmer, Sven Lehmann, Rainer Sch체ssler
}
