% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{inflation_data}
\alias{inflation_data}
\title{Data to estimate quarterly U.S. inflation}
\format{
A \link{data.frame} with 248 quarterly observations (rows) and 445 variables (columns).
}
\source{
\url{https://onlinelibrary.wiley.com/doi/full/10.1111/iere.12623}
}
\usage{
inflation_data
}
\description{
A novel, high-dimensional dataset built by Koop and Korobilis (2023)
              that merges predictors from several mainstream aggregate
              macroeconomic and financial data sets. The dataset includes
              the FRED-QD dataset of McCracken and Ng (2020),
              augment with portfolio data used in Jurado et al. (2015),
              stock market predictors from Welch and Goyal (2008),
              survey data from University of Michigan consumer surveys,
              commodity prices from the World Bank’s Pink Sheet database,
              and key macroeconomic indicators from the Federal Reserve Economic Data
              for four economies (Canada, Germany, Japan, United Kingdom).
              All data are quarterly, and span the period 1960Q1–2021Q4.
}
\references{
Jurado, K., Ludvigson, S. C., and Ng, S. (2015) "Measuring uncertainty."
\emph{American Economic Review}, 105 (3): 1177–1216.

Koop, G. and Korobilis, D. (2023) "Bayesian dynamic variable selection in high dimensions."
\emph{International Economic Review}.

McCracken, M., and S. Ng (2020) “FRED-QD: A Quarterly Database for Macroeconomic Research”
\emph{National Bureau of Economic Research}, Working Paper 26872.

Welch, I. and Goyal, A. (2008) "A comprehensive look at the empirical performance of equity premium prediction."
\emph{The Review of Financial Studies}, 21 (4): 1455–1508.
}
\keyword{datasets}
