% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassMethods.R
\name{summary.dsc_obj}
\alias{summary.dsc_obj}
\title{Plots the evolution of the tuning parameter for 'dsc' object
and returns basic performance metrics}
\usage{
\method{summary}{dsc_obj}(object, eval_period = NULL, ...)
}
\arguments{
\item{object}{An object of type 'dsc'}

\item{eval_period}{(Optional) A vector of indices to specify the evaluation period}

\item{...}{Additional arguments to be consistent with S3 print() function}
}
\description{
Plots the evolution of the tuning parameter for 'dsc' object
and returns basic performance metrics
}
\references{
Gneiting, T., Raftery, A. E., Westveld, A. H., and Goldman, T. (2005): Calibrated Probabilistic Forecasting Using Ensemble Model Output Statistics and Minimum CRPS Estimation. \emph{Monthly Weather Review}, 133: 1098–1118.
Jordan, A., Krueger, F., and Lerch, S. (2019): "Evaluating Probabilistic Forecasts with scoringRules." \emph{Journal of Statistical Software}, 90(12): 1-37.
}
