% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsc.R
\name{dsc}
\alias{dsc}
\title{Generate Dynamic Subset Forecast Combinations}
\usage{
dsc(gamma_grid, psi_grid, ew_grid, y, mu_mat, var_mat, delta, n_cores)
}
\arguments{
\item{gamma_grid}{Numerical Vector.
Tuning parameter to exponentially down-weight past predictive performance.}

\item{psi_grid}{Integer Vector.
Tuning parameter that controls the size of the active subsets
(number of predictive densities that are combined).}

\item{ew_grid}{Boolean.
Choose 1 to assign equal weights (that sum to one)
to an active subset of forecasting models.
Choose 0 to assign the weights based on the corresponding
predictive likelihoods in t - 1 to an active subset of forecasting models.}

\item{y}{A matrix of dimension `T * 1` or Numeric Vector of length `T`
containing the observations of the target variable.}

\item{mu_mat}{A matrix with `T` rows containing
the point forecasts of each predictive density in each column.}

\item{var_mat}{A matrix with `T` rows containing
the variance of each predictive density in each column.}

\item{delta}{Double. Value denoting the discount factor used
to down-weight past predictive performance.}

\item{n_cores}{Integer. The number of cores to use for the estimation.}
}
\value{
List that contains
(1) the point forecast (2) the variance of the Dynamic Subset Combination
and (3) a list that contains the column-index of the
selected density forecasts.
}
\description{
`dsc()` can be used to generate forecast combinations
from a set of candidate density forecasts. For each period,
`dsc()` selects a subset of predictive densities with highest ranks
regarding local predictive accuracy.
By using a discount factor gamma,
past predictive performance is exponentially down-weighted.
Both the identities of the forecasting models
that are used for building the combined forecast and
the subset sizes may vary over time based on the data.
If only one forecasting model is picked, the approach (temporarily)
collapses to pure model selection.
}
