% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_dsc.R
\name{eval_dsc}
\alias{eval_dsc}
\title{Evaluate Dynamic Subset Combination Results}
\usage{
eval_dsc(
  oos_y,
  oos_benchmark,
  oos_forecast_dsc,
  oos_dates,
  oos_chosen_parameter,
  model_names_tvc,
  oos_models_idx
)
}
\arguments{
\item{oos_y}{y A matrix of dimension `T * 1` or Numeric Vector of length `T`
containing the out-of-sample observations of the target variable.}

\item{oos_benchmark}{A matrix of dimension `T * 1` or
Numeric Vector of length `T` containing the
out-of-sample forecasts of a arbitrary benchmark.}

\item{oos_forecast_dsc}{A matrix of dimension `T * 1`
or Numeric Vector of length `T` containing the
out-of-sample forecasts of the Dynamic Subset Combination.}

\item{oos_dates}{A vector that contains all ouf-of-sample dates.}

\item{oos_chosen_parameter}{Matrix from 'dsc' that contains the
chosen (tuned) parameter from the Dynamic Subset Combination.}

\item{model_names_tvc}{Character Vector from [tvp()] that contains
all model names from the density forecasts.}

\item{oos_models_idx}{List from [dsc()] that contains
the column-indices of the selected density forecasts.}
}
\value{
List that contains
(1) the Clark-West-Statistic,
(2) the Out-of-Sample R2,
(3) the CSED between the DSC-Forecast and the benchmark,
(4) a boolean matrix that indicates when a predictor was chosen from the DSC
(5) and a CSED Plot.
}
\description{
`eval_dsc()` can be used to evaluate the results from [dsc()].
It provides statistical measures such as Clark-West-Statistic or the OOS-R2,
visualizes for the predictive performance as well as the selected parameters
and returns a matrix which indicates the selected predictors
at every point in time.
}
\examples{
\donttest{

  ### Simulate Data
  set.seed(123)

  # Set Dimensions
  numb_obs   <-  500
  numb_pred  <-  50
  numb_forc  <-  10

  # Create Random Target-Variable
  equity_premium  <-  rnorm(n = numb_obs, mean = 0, sd = 1)

  # Create Random Text-Variables
  raw_preds            <-  replicate(numb_pred, sample(0:10,
                                                       numb_obs,
                                                       rep = TRUE), )
  raw_names            <-  paste0("X", as.character(seq_len(numb_pred)))
  colnames(raw_preds)  <-  raw_names

  # Create Random Point Forecasts
  f_preds            <-  replicate(10, rnorm(n    = numb_obs,
                                             mean = 0,
                                             sd   = 0.5), )
  f_names            <-  paste0("F", as.character(seq_len(numb_forc)))
  colnames(f_preds)  <-  f_names

  # Create Benchmark
  benchmark  <-  dplyr::lag(roll::roll_mean(equity_premium,
                                            width = length(equity_premium),
                                            min_obs = 1), n = 1)

  # Specify TV-C-Parameter
  sample_length  <-  floor(numb_obs / 10)
  lambda_grid    <-  c(0.9995, 0.9999, 1.0000)
  kappa_grid     <-  c(0.94)
  n_cores        <-  1

  # Apply TV-C-Function
  results  <-  hdflex::tvc(equity_premium,
                           raw_preds,
                           f_preds,
                           lambda_grid,
                           kappa_grid,
                           sample_length,
                           n_cores)

  # Assign Results
  forecast_tvc      <-  results[[1]]
  variance_tvc      <-  results[[2]]
  model_names_tvc   <-  results[[3]]

  # Cut Initialization-Period
  nr_drp              <-  dim(forecast_tvc)[1] -
                          dim(na.omit(forecast_tvc))[1] + 1
  sample_period_idx   <-  (nr_drp + sample_length):numb_obs

  # Trim Objects
  sub_forecast_tvc    <-  forecast_tvc[sample_period_idx, , drop = FALSE]
  sub_variance_tvc    <-  variance_tvc[sample_period_idx, , drop = FALSE]
  sub_benchmark       <-  benchmark[sample_period_idx]
  sub_equity_premium  <-  equity_premium[sample_period_idx]

  ##### Dynamic Subset Combination #####
  # Set DSC Parameter
  nr_mods     <-  length(model_names_tvc)
  gamma_grid  <-  c(0.9, 0.95, 0.99, 1)
  psi_grid    <-  c(1, 2, 3, 4, 5)
  delta       <-  0.9992
  n_cores     <-  1

  # Apply DSC-Function
  results  <-  hdflex::dsc(gamma_grid,
                           psi_grid,
                           sub_equity_premium,
                           sub_forecast_tvc,
                           sub_variance_tvc,
                           delta,
                           n_cores)

  # Assign Results
  sub_forecast_dsc    <-  results[[1]]
  sub_variance_dsc    <-  results[[2]]
  model_names_comb    <-  results[[3]]
  sub_chosen_para     <-  results[[4]]
  sub_models_idx      <-  results[[5]]

  # Define Evaluation Period
  eval_period_idx     <-  50:length(sub_equity_premium)

  # Trim Objects
  oos_equity_premium  <-  sub_equity_premium[eval_period_idx]
  oos_benchmark       <-  sub_benchmark[eval_period_idx]
  oos_forecast_dsc    <-  sub_forecast_dsc[eval_period_idx]
  oos_variance_dsc    <-  sub_variance_dsc[eval_period_idx]
  oos_models_idx      <-  lapply(seq_along(model_names_comb), function(i) {
                                      sub_models_idx[[i]][eval_period_idx]})
  oos_chosen_para     <-  sub_chosen_para[eval_period_idx, , drop = FALSE]
  oos_dates           <-  seq(as.Date("1989-12-01"),
                              by = "day",
                              length.out = length(eval_period_idx))

  # Assign Names
  names(oos_forecast_dsc)  <-  oos_dates
  names(oos_variance_dsc)  <-  oos_dates

  # Apply Statistial-Evaluation-Function
  eval_results  <-  eval_dsc(oos_equity_premium,
                             oos_benchmark,
                             oos_forecast_dsc,
                             oos_dates,
                             oos_chosen_para,
                             model_names_tvc,
                             oos_models_idx)

  # Assign Results
  cw_t          <-  eval_results[[1]]
  oos_r2        <-  eval_results[[2]]
  csed          <-  eval_results[[3]]
  pred_pockets  <-  eval_results[[4]]
}
}
