% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tvc.R
\name{tvc}
\alias{tvc}
\title{Compute Univariate Time-Varying Coefficient (TV-C) Regressions}
\usage{
tvc(y, x, f, lambda_grid, kappa_grid, init_length, x_names, f_names, n_cores)
}
\arguments{
\item{y}{A matrix of dimension `T * 1` or Numeric Vector of length `T`
containing the observations of the target variable.}

\item{x}{A matrix with `T` rows containing
the lagged raw predictors in each column.
Use `NULL` if no "raw" predictors shall be included.}

\item{f}{A matrix with `T` rows containing
point forecasts of y in each column.
Use NULL if no point forecasts shall be included.}

\item{lambda_grid}{A numeric vector denoting the discount factor(s)
that control the dynamics of the coefficients.
Each predictor in combination with each value of
lambda defines a separate forecasting model.
Constant coefficients are nested for the case `lambda = 1`.}

\item{kappa_grid}{A numeric vector to accomodate time-varying volatility.
The observational variance is estimated via
Exponentially Weighted Moving Average.
Constant variance is nested for the case `kappa = 1`.
Each predictor in combination with each value of
kappa defines a separate forecasting model.}

\item{init_length}{Integer. Denotes the number of observations used
to initialize (estimate) the the TV-C models
(e.g. the observational variance).}

\item{x_names}{Character vector with all 'raw' predictor names.
Use NULL if no 'raw' predictors are included.}

\item{f_names}{Character vector with all point forecast names.
Use NULL if no point forecasts are included.}

\item{n_cores}{Integer. The number of cores to use for the estimation.}
}
\value{
List that contains (1) a matrix with all forecasts,
(2) a matrix with all variances and (3) a vector with all model names.
}
\description{
`tvc()` can be used to generate density forecasts based on
univariate time-varying coefficient regressions
(nesting constant coefficients as a special case)
with time-varying volatility. In addition to “raw” predictors,
`tvc()` accommodates point forecast as predictors.
}
